scrap/requirements.txt
fastapi==0.110.0
scrapy==2.11.0        
pydantic==2.6.4
python-dotenv==1.0.1
pymongo==4.6.3
beautifulsoup4==4.12.3 
